protein-rna-viewer/pyproject.toml
[project]
name = "bio-pipeline"
version = "0.1.0"
dependencies = [
    "biopython",
    "rich",
    "python-dotenv",
]

[project.optional-dependencies]
rna = ["ViennaRNA"]